moved the reflection function to a separate file
