+++ lesson7/reflection/reflecation.go
@@ -1,20 +1,9 @@
-//Go. Уровень 2. Урок 7. Рефлексия и кодогенерация.
-//Написать функцию, которая принимает на вход структуру in (struct или кастомную struct)
-//и values map[string]interface{} (key - название поля структуры, которому нужно присвоить value этой мапы).
-//Необходимо по значениям из мапы изменить входящую структуру in с помощью пакета reflect.
-//Функция может возвращать только ошибку error.
-//Написать к данной функции тесты (чем больше, тем лучше - зачтется в плюс).
-//основная цель рефлексии - это определение типа и значения, хранящегося в объекте интерфейса
-package main
+package reflection
 
 import (
 	"fmt"
-	"log"
+	"reflect"
 )
-
-//type appropriator interface {
-//	reflectionExample(m map[string]interface{}) error
-//}
 
 type In struct {
 	Key int
@@ -22,16 +11,15 @@ type In struct {
 
 var Structure In
 
-/*
 // Reflection функция заполняет поля структуры значениями из карты, возвращает только ошибку error.
-func Reflection(m map[string]interface{}, i *In) error {
+func Reflection(m map[string]interface{}, Structure In) error {
 	//fmt.Println(reflect.TypeOf(m))
 	for k, v := range m {
 		//fmt.Println("Key:", k, "Value:", v)
 		//fmt.Println(reflect.TypeOf(k))
-		i.Key = int(reflect.ValueOf(v).Int())
+		Structure.Key = int(reflect.ValueOf(v).Int())
 		//i.Key = v.(int) - это работает, но не использует reflect
-		fmt.Println(i.Key, i)
+		fmt.Println(Structure.Key, Structure)
 
 		t := reflect.ValueOf(k)                       //получаем значение ключа карты
 		fmt.Println(t, "have type of key:", t.Type()) //смотрим тип значения ключа карты
@@ -51,6 +39,7 @@ func Reflection(m map[string]interface{}, i *In) error {
 			case bool:
 				return fmt.Errorf("type mismatch, expected value as string")
 			}
+		*/
 
 		//проверка соответствия типа значения элемента карты (мапы)
 		switch v.(type) {
@@ -65,20 +54,4 @@ func Reflection(m map[string]interface{}, i *In) error {
 		}
 	}
 	return nil
-}
-*/
-func main() {
-	ValuesMap := map[string]interface{}{
-		"rsc": 3711,
-		"r":   2138,
-		"gri": 1908,
-		"adg": 912,
-	}
-	fmt.Printf("Begin struct structure %v\n", Structure)
-	err := Reflection(ValuesMap, &Structure)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Printf("Finish struct structure %v\n", Structure)
-
 }
